"In this version it was created a file called deviance_estimator.m to perform the MLE calculations.
Also, this version is no longer using the ESPRIT, SPS and FBA source codes used to perform the Eduardos' TCC measures. Now we are using Daniel's source code. This change did not make any difference. I am using Daniel's souce code because the it is cleaner and does not use tensor approach.

"